allow hotswapping db configs and separate server from exports
