+++ ts/src/db_connection.ts
@@ -1,5 +1,9 @@
 import * as _ from 'lodash';
-import { Connection, createConnection } from 'typeorm';
+import { Connection, ConnectionOptions, createConnection } from 'typeorm';
+
+import { OrderEntity } from './entities/order_entity';
+import { TakerAssetFillAmountEntity } from './entities/taker_asset_fill_amount_entity';
+import { TransactionEntity } from './entities/transaction_entity';
 
 let connectionIfExists: Connection | undefined;
 
@@ -24,9 +28,22 @@ export function getDBConnection(): Connection {
 /**
  * Creates the DB connnection to use in an app
  */
-export async function initDBConnectionAsync(): Promise<void> {
+export async function initDBConnectionAsync(options?: ConnectionOptions): Promise<void> {
     if (!_.isUndefined(connectionIfExists)) {
         throw new Error('DB connection already exists');
     }
-    connectionIfExists = await createConnection();
+    let connOptions = options;
+    if (connOptions === undefined) {
+        connOptions = {
+            type: 'sqlite',
+            database: 'database.sqlite',
+            synchronize: true,
+            logging: true,
+            entities: [OrderEntity, TakerAssetFillAmountEntity, TransactionEntity],
+            cli: {
+                entitiesDir: './entities',
+            },
+        };
+    }
+    connectionIfExists = await createConnection(connOptions);
 }
